Message Service soweit fertig

+++ server/src/main/java/service/interfaces/MessageService.java
@@ -6,9 +6,9 @@ import java.util.ArrayList;
 
 public interface MessageService {
     void addMesssage(int groupId, int userId, String content);
-    void removeMessage(int id);
-    Message getMessageById(int id);
-    String getMessageContentById(int id);
+    void removeMessage(long id);
+    Message getMessageById(long id);
+    String getMessageContentById(long id);
     ArrayList<Message> getMessagesForUser(int userId);
     ArrayList<Message> getMessagesForGroup(int groupId);
